Merge fixes from upstream
This pulls in the list and rm fixes from the upstream
j-griffith.  We also go ahead and increment the version
being reported by the binary so now we can update our
binary in the repo.

We're probably pretty much ready to make this the primary repo
instead of a fork/downstream repo.